+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog - LOSPOR Core
 
-## [Unreleased]
+## [9.10.0] - 2026-09-15
 
 ### Added